+++ src/index.ts
@@ -6,17 +6,18 @@ export class html2mrkdwn {
      * @hideconstructor
      */
     constructor() {
-        
+
     };
-    
+
     /**
      * Rudimentary method for converting HTML code to mrkdown
-     * 
+     *
      * @static
      * @param html HTML code
+     * @param removeOtherTags Remove all tags that are not supported by Slack Mrkdwn
      * @returns converted HTML to Slack Mrkdwn
      */
-    public static convert(html: string): string {
+    public static convert(html: string, removeOtherTags: boolean): string {
 
         let converted:string = "";
 
@@ -62,9 +63,22 @@ export class html2mrkdwn {
         // Parse Links
         converted = converted.replace(/<a href="(.*)">(.*)<\/a>/g, "<$1|$2>");
 
+        if(removeOtherTags) {
+            // Remove all other tags
+            converted = converted.replace(/<[^>]+>/g, "");
+        }
 
 
         return converted;
+    }
+
+    public static extractImg(html: string): string[] {
+        const imgTags = html.match(/<img [^>]*src="[^"]*"[^>]*>/g) || [];
+        const imgSrcs = imgTags.map(tag => {
+            const match = tag.match(/src="([^"]*)"/);
+            return match ? match[1] : "";
+        });
+        return imgSrcs;
     }
 
 }