+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## [next]
+## [0.2.2]
+
+### Added
+
+- add bool for removing tags not supported by mrkdwn
+- add method to extract the image src link
+- updated dependencies
+
+### Removed
+
+- npm link script
 
 ## [0.1.0] - 2023-11-27